fix copilot comment 1 and 2
